pyproject.toml
[tool.poetry]
name = "discordrpg"
version = "0.1.0"
description = "An RPG bot for discord..."
authors = ["TimTam <notgivingyou@my.email>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.14.0"
django-user = "^0.2.1"
objectpath = "^0.6.1"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
